+++ src/interface/mouse-handler.js
